+++ foundry/settings_dev.py
@@ -1,9 +1,11 @@
 from foundry.settings import *
 
 #FOUNDRY['layers'] =  ('basic',)
+#FOUNDRY['layers'] =  ('smart', 'basic',)
 FOUNDRY['layers'] =  ('web', 'basic',)
-#FOUNDRY['layers'] =  ('mid', 'basic',)
-#FOUNDRY['layers'] =  ('smart', 'mid', 'basic',)
 
-
-compute_settings(sys.modules[__name__])
+CACHES = {
+    'default': {
+        'BACKEND': 'django.core.cache.backends.dummy.DummyCache',
+    }
+}